+++ mcp/packages/server/src/PenpotMcpServer.ts
@@ -21,6 +21,7 @@ import {
 } from "./tools/GlobalReadTools";
 import { FileCreateTool } from "./tools/FileCreateTool";
 import { FileBindContextTool, FileGetContextTool, FileReleaseContextTool } from "./tools/FileContextTools";
+import { PageCreateTool, PageListTool, PageRenameTool, PageSetCurrentTool } from "./tools/PageTools";
 import { PenpotRpcClient } from "./PenpotRpcClient";
 import { ReplServer } from "./ReplServer";
 import { ApiDocs } from "./ApiDocs";
@@ -215,6 +216,10 @@ export class PenpotMcpServer {
             new FileGetContextTool(this),
             new FileBindContextTool(this),
             new FileReleaseContextTool(this),
+            new PageListTool(this),
+            new PageCreateTool(this),
+            new PageRenameTool(this),
+            new PageSetCurrentTool(this),
             new ExecuteCodeTool(this),
             new HighLevelOverviewTool(this),
             new PenpotApiInfoTool(this, this.apiDocs),